--- backend/src/main/java/com/prograngers/backend/dto/comment/CommentPatchRequest.java
@@ -1,17 +1,0 @@
-package com.prograngers.backend.dto.comment;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class CommentPatchRequest {
-    @NotBlank(message = "댓글 내용을 입력해주세요")
-    private String content;
-    private String mention;
-}
