+++ backend/src/main/java/com/prograngers/backend/dto/comment/request/CommentReqeust.java
@@ -1,6 +1,6 @@
-package com.prograngers.backend.dto.solution;
+package com.prograngers.backend.dto.comment.request;
 
-
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,10 +10,17 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class SolutionDetailResponseComment {
+public class CommentReqeust {
+
     String nickname;
     Integer orderParent;
-    Integer orderChild;
+    @NotBlank(message = "댓글 내용을 입력해주세요")
     String content;
+
     String mention;
+
+    Long parentId;
+
+    Integer groupNumber;
+
 }
