+++ backend/src/main/java/com/prograngers/backend/dto/review/response/ReviewResponse.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.review;
+package com.prograngers.backend.dto.review.response;
 
 import com.prograngers.backend.entity.Review;
 import lombok.Builder;
