+++ backend/src/main/java/com/prograngers/backend/dto/solution/reqeust/SolutionPatchRequest.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.solution;
+package com.prograngers.backend.dto.solution.reqeust;
 
 import com.prograngers.backend.entity.*;
 import com.prograngers.backend.entity.constants.AlgorithmConstant;
