+++ backend/src/main/java/com/prograngers/backend/dto/solution/response/SolutionDetailResponse.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.solution;
+package com.prograngers.backend.dto.solution.response;
 
 import com.prograngers.backend.entity.Comment;
 import com.prograngers.backend.entity.Solution;
