+++ backend/src/main/java/com/prograngers/backend/dto/solution/response/SolutionResponse.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.dto.solution;
+package com.prograngers.backend.dto.solution.response;
 
 import com.prograngers.backend.entity.constants.AlgorithmConstant;
 import com.prograngers.backend.entity.constants.DataStructureConstant;
