+++ backend/src/main/java/com/prograngers/backend/service/ProblemService.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.service;
 
-import com.prograngers.backend.dto.problem.ProblemResponse;
+import com.prograngers.backend.dto.problem.response.ProblemResponse;
 import com.prograngers.backend.entity.Problem;
 import com.prograngers.backend.entity.constants.AlgorithmConstant;
 import com.prograngers.backend.entity.constants.DataStructureConstant;
